in Heroku guide, updated Heroku install instructions
I replaced the first set of instructions with a link, so the instructions are less brittle. They currently say “gem install heroku”, but Heroku now recommends that you install the Heroku Toolbelt and makes no mention of the gem. The instructions could change again, so I now only summarize the steps. I thought “installing the Heroku client” was slightly better than “installing the Heroku Toolbelt” because it was less brittle, even though it’s a little less accurate.

I copied my added description of the Heroku quick start steps to the similar link in “Step 0” of the following section “Deploying an existing local Refinery application”.