Generator specs clean up after themselves.
I was having issues running the generator specs because I got permission
denied errors for the tmpdir that was being used in the spec for
engine_generator_multiple_resources.

I then noticed that the other specs use a relative path inside the core/
directory for the temporary directory.

I've switched the spec to do this and also added an after which cleans
up after the test ensuring no extra files or folders are left lying
around on the file system.